requirements.txt
django==3.1.2
djangorestframework==3.12.1
sqlparse==0.3.1
pytz==2020.1
django-cors-headers==3.5.0
boto3==1.15.15
requests==2.25.0
mfrc522==0.0.7
RPi.GPIO==0.7.0
picamera==1.13
